Working on a Factory
I am attempting to create a factory contract that keeps track of individual SBTs.  Not sure if I need to create a second contract to achieve this.  Also having difficulties deploying to hardhat.  I think this is due to now having two contracts.  Created second branch to start this idea.  I may abandon this idea if I can keep everything on one contract.